extract FeaturesSet as its own concern
