+++ lib/iodized2_ruby_client/client.rb
@@ -1,12 +1,15 @@
 require 'iodized2_ruby_client/ws_client'
+require 'iodized2_ruby_client/features_set'
 
 module Iodized2RubyClient
   class Client
-    attr_reader :features
+    extend Forwardable
+
+    def_delegators :@features, :features, :enabled?
 
     def initialize(url)
       @url = url
-      @features = []
+      @features = FeaturesSet.new
 
       @ws_client = Iodized2RubyClient::WSClient.new(@url) do |message|
         handle_message(message)
@@ -20,32 +23,22 @@ module Iodized2RubyClient
       send("handle_#{key}", result[key])
     end
 
-    def enabled?(feature_name)
-      feature = @features.find { |feature| feature["name"] == feature_name }
-      feature["active"] if feature
-    end
-
     private
 
     def handle_sync(features)
-      @features = features.freeze
+      @features.sync_features(features)
     end
 
     def handle_create(feature)
-      @features = (@features.dup << feature).freeze
+      @features.add_feature(feature)
     end
 
-    # this is essentially a delete, followed by a create
-    # but if we simply called them in sequence we would leave a point where a
-    # valid feature momentarily disappears from the set of features.
-    # so to avoid needing mutexes around each access, we will effectively duplicate
-    # those two operations into a single update.
     def handle_update(feature)
-      @features = (@features.reject { |f| f["id"] == feature["id"] } << feature).freeze
+      @features.update_feature(feature)
     end
 
     def handle_delete(feature)
-      @features = @features.reject { |f| f["id"] == feature["id"] }.freeze
+      @features.delete_feature(feature)
     end
   end
 end
